move checkbox to packages
