+++ packages/@guardian/src-checkbox/Checkbox.stories.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
-import type { Story } from '../../../../lib/@types/storybook-emotion-10-fixes';
-import { asPlayground, asChromaticStory } from '../../../../lib/story-intents';
+import type { Story } from '../../../lib/@types/storybook-emotion-10-fixes';
+import { asPlayground, asChromaticStory } from '../../../lib/story-intents';
 // These types are the right types, but don't work with Storybook v6 which uses Emotion v10
 // import type { Args, Story } from '@storybook/react';
 import { Checkbox } from './Checkbox';
