--- src/core/components/checkbox/tsconfig.json
@@ -1,24 +1,0 @@
-{
-  "extends": "../../../../tsconfig.json",
-  "compilerOptions": {
-    "outDir": "dist/types",
-    "declaration": true,
-    "emitDeclarationOnly": true
-  },
-  "include": [".", "../../foundations/types/themes.d.ts"],
-  "exclude": ["dist", "*.stories.tsx"],
-  "references": [
-    {
-      "path": "../../foundations"
-    },
-    {
-      "path": "../../helpers"
-    },
-    {
-      "path": "../user-feedback"
-    },
-    {
-      "path": "../label"
-    }
-  ]
-}
